pyproject.toml
[tool.poetry]
name = "ai-app"
version = "0.1.0"
description = ""
authors = ["dany1940 <59851426+dany1940@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.30"
pydantic = "^2.7.1"
psycopg2 = "^2.9.9"
alembic = "^1.13.1"
sqlmodel = "^0.0.18"
uvicorn = "^0.30.1"
fastapi = "^0.111.0"
fastapi-health = "^0.4.0"
python-dotenv = "^1.0.1"
gunicorn = "^22.0.0"
pytest = "^8.2.2"
isort = "^5.13.2"
black = "^24.4.2"
mypy = "^1.10.0"
types-psycopg2 = "^2.9.21.20240417"
sqlalchemy-utils = "^0.41.2"
python-jose = "^3.3.0"
python-jwt = "^4.1.0"
pyjwt = "2.1.0"
passlib = "^1.7.4"
bcrypt = "^4.1.3"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
